Expose Sqlite 3.24 test-fixtures
