+++ dialects/sqlite-3-24/build.gradle
@@ -2,6 +2,7 @@ plugins {
   alias(libs.plugins.kotlin.jvm)
   alias(libs.plugins.grammarKitComposer)
   alias(libs.plugins.publish)
+  alias(libs.plugins.dokka)
   id("app.cash.sqldelight.toolchain.compiler")
   id("java-test-fixtures")
 }
@@ -17,14 +18,13 @@ dependencies {
     targetConfiguration = "shadow"
   }
 
-  testImplementation(projects.sqldelightCompiler.environment) {
+  testFixturesApi(projects.sqldelightCompiler.environment) {
     targetConfiguration = "shadow"
   }
+  testFixturesApi testFixtures(libs.sqlPsi)
+  testFixturesApi testFixtures(projects.dialects.sqlite318)
+
   testImplementation libs.junit
-  testImplementation libs.truth
-  testImplementation(testFixtures(libs.sqlPsi)) {
-    exclude group: "com.jetbrains.intellij.platform"
-  }
 }
 
 apply from: "$rootDir/gradle/gradle-mvn-push.gradle"
