+++ dialects/sqlite-3-24/src/test/kotlin/app/cash/sqldelight/dialects/sqlite_3_24/Sqlite324FixturesTest.kt
@@ -5,6 +5,7 @@ import java.io.File
 import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
 import org.junit.runners.Parameterized.Parameters
+import app.cash.sqldelight.dialects.sqlite_3_18.SqliteTestFixtures as Sqlite_3_18SqliteTestFixtures
 
 @RunWith(Parameterized::class)
 class Sqlite324FixturesTest(name: String, fixtureRoot: File) : FixturesTest(name, fixtureRoot) {
@@ -17,16 +18,10 @@ class Sqlite324FixturesTest(name: String, fixtureRoot: File) : FixturesTest(name
   }
 
   companion object {
-    private val fixtures = arrayOf("../sqlite-3-18/src/test/fixtures_sqlite_3_18", "src/test/fixtures_sqlite_3_24")
-
     @Suppress("unused")
     // Used by Parameterized JUnit runner reflectively.
     @Parameters(name = "{0}")
     @JvmStatic
-    fun parameters() = fixtures.flatMap { fixtureFolder ->
-      File(fixtureFolder).listFiles()!!
-        .filter { it.isDirectory }
-        .map { arrayOf(it.name, it) }
-    } + ansiFixtures
+    fun parameters() = Sqlite_3_18SqliteTestFixtures.fixtures_sqlite_3_18 + SqliteTestFixtures.fixtures + ansiFixtures
   }
 }
