+++ dialects/sqlite-3-24/src/testFixtures/resources/fixtures_sqlite_3_24/extension_expr/Test.s
